requirements.txt
allure-pytest==2.5.1
allure-python-commons==2.5.1
apipkg==1.5
atomicwrites==1.1.5
atpublic==1.0
attrs==18.1.0
certifi==2018.8.13
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
Django==2.1
execnet==1.5.0
idna==2.7
kiwisolver==1.0.1
matplotlib==2.2.3
more-itertools==4.3.0
numpy==1.15.1
pluggy==0.7.1
py==1.5.4
pyparsing==2.2.0
pytest==3.7.2
pytest-base-url==1.4.1
pytest-forked==0.2
pytest-html==1.19.0
pytest-metadata==1.7.0
pytest-rerunfailures==4.1
pytest-selenium==1.13.0
pytest-variables==1.7.1
pytest-xdist==1.22.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
selenium==3.14.0
six==1.11.0
urllib3==1.23
